+++ libraries/exoplayer/src/test/java/androidx/media3/exoplayer/ExoPlayerTest.java
@@ -171,6 +171,7 @@ import androidx.media3.test.utils.robolectric.TestPlayerRunHelper;
 import androidx.test.core.app.ApplicationProvider;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import com.google.common.collect.ImmutableList;
+import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Iterables;
 import com.google.common.collect.Range;
 import com.google.errorprone.annotations.CanIgnoreReturnValue;
@@ -5050,25 +5051,40 @@ public final class ExoPlayerTest {
     ArgumentCaptor<PositionInfo> newPositionArgumentCaptor =
         ArgumentCaptor.forClass(PositionInfo.class);
     ArgumentCaptor<Integer> reasonArgumentCaptor = ArgumentCaptor.forClass(Integer.class);
-    FakeTimeline adTimeline =
+    // Create a multi-period timeline without ads.
+    FakeTimeline fakeContentTimeline =
+        new FakeTimeline(
+            new FakeTimeline.TimelineWindowDefinition(
+                /* periodCount= */ 4,
+                "windowId",
+                /* isSeekable= */ true,
+                /* isDynamic= */ false,
+                /* isLive= */ false,
+                /* isPlaceholder= */ false,
+                /* durationUs= */ DEFAULT_WINDOW_DURATION_US,
+                /* defaultPositionUs= */ 0,
+                /* windowOffsetInFirstPeriodUs= */ DEFAULT_WINDOW_OFFSET_IN_FIRST_PERIOD_US,
+                /* adPlaybackStates= */ ImmutableList.of(AdPlaybackState.NONE),
+                MediaItem.EMPTY));
+    // Create the ad playback state matching to the periods in the content timeline.
+    ImmutableMap<Object, AdPlaybackState> adPlaybackStates =
         FakeTimeline.createMultiPeriodAdTimeline(
-            "windowId",
-            /* numberOfPlayedAds= */ 0,
-            /* isAdPeriodFlags...= */ false,
-            true,
-            true,
-            false);
+                "windowId",
+                /* numberOfPlayedAds= */ 0,
+                /* isAdPeriodFlags...= */ false,
+                true,
+                true,
+                false)
+            .getAdPlaybackStates(/* windowIndex= */ 0);
     Listener listener = mock(Listener.class);
     ExoPlayer player = new TestExoPlayerBuilder(context).build();
     player.addListener(listener);
     AtomicReference<ServerSideAdInsertionMediaSource> sourceReference = new AtomicReference<>();
     sourceReference.set(
         new ServerSideAdInsertionMediaSource(
-            new FakeMediaSource(adTimeline),
+            new FakeMediaSource(fakeContentTimeline),
             contentTimeline -> {
-              sourceReference
-                  .get()
-                  .setAdPlaybackStates(adTimeline.getAdPlaybackStates(/* windowIndex= */ 0));
+              sourceReference.get().setAdPlaybackStates(adPlaybackStates, contentTimeline);
               return true;
             }));
     player.setMediaSource(sourceReference.get());
@@ -5141,25 +5157,40 @@ public final class ExoPlayerTest {
     ArgumentCaptor<PositionInfo> newPositionArgumentCaptor =
         ArgumentCaptor.forClass(PositionInfo.class);
     ArgumentCaptor<Integer> reasonArgumentCaptor = ArgumentCaptor.forClass(Integer.class);
-    FakeTimeline adTimeline =
+    // Create a multi-period timeline without ads.
+    FakeTimeline fakeContentTimeline =
+        new FakeTimeline(
+            new FakeTimeline.TimelineWindowDefinition(
+                /* periodCount= */ 4,
+                "windowId",
+                /* isSeekable= */ true,
+                /* isDynamic= */ false,
+                /* isLive= */ false,
+                /* isPlaceholder= */ false,
+                /* durationUs= */ DEFAULT_WINDOW_DURATION_US,
+                /* defaultPositionUs= */ 0,
+                /* windowOffsetInFirstPeriodUs= */ DEFAULT_WINDOW_OFFSET_IN_FIRST_PERIOD_US,
+                /* adPlaybackStates= */ ImmutableList.of(AdPlaybackState.NONE),
+                MediaItem.EMPTY));
+    // Create the ad playback state matching to the periods in the content timeline.
+    ImmutableMap<Object, AdPlaybackState> adPlaybackStates =
         FakeTimeline.createMultiPeriodAdTimeline(
-            "windowId",
-            /* numberOfPlayedAds= */ 0,
-            /* isAdPeriodFlags...= */ false,
-            true,
-            false,
-            false);
+                "windowId",
+                /* numberOfPlayedAds= */ 0,
+                /* isAdPeriodFlags...= */ false,
+                true,
+                false,
+                false)
+            .getAdPlaybackStates(/* windowIndex= */ 0);
     Listener listener = mock(Listener.class);
     ExoPlayer player = new TestExoPlayerBuilder(context).build();
     player.addListener(listener);
     AtomicReference<ServerSideAdInsertionMediaSource> sourceReference = new AtomicReference<>();
     sourceReference.set(
         new ServerSideAdInsertionMediaSource(
-            new FakeMediaSource(adTimeline),
+            new FakeMediaSource(fakeContentTimeline),
             contentTimeline -> {
-              sourceReference
-                  .get()
-                  .setAdPlaybackStates(adTimeline.getAdPlaybackStates(/* windowIndex= */ 0));
+              sourceReference.get().setAdPlaybackStates(adPlaybackStates, contentTimeline);
               return true;
             }));
     player.setMediaSource(sourceReference.get());
@@ -5203,25 +5234,40 @@ public final class ExoPlayerTest {
     ArgumentCaptor<PositionInfo> newPositionArgumentCaptor =
         ArgumentCaptor.forClass(PositionInfo.class);
     ArgumentCaptor<Integer> reasonArgumentCaptor = ArgumentCaptor.forClass(Integer.class);
-    FakeTimeline adTimeline =
+    // Create a multi-period timeline without ads.
+    FakeTimeline fakeContentTimeline =
+        new FakeTimeline(
+            new FakeTimeline.TimelineWindowDefinition(
+                /* periodCount= */ 4,
+                "windowId",
+                /* isSeekable= */ true,
+                /* isDynamic= */ false,
+                /* isLive= */ false,
+                /* isPlaceholder= */ false,
+                /* durationUs= */ DEFAULT_WINDOW_DURATION_US,
+                /* defaultPositionUs= */ 0,
+                /* windowOffsetInFirstPeriodUs= */ DEFAULT_WINDOW_OFFSET_IN_FIRST_PERIOD_US,
+                /* adPlaybackStates= */ ImmutableList.of(AdPlaybackState.NONE),
+                MediaItem.EMPTY));
+    // Create the ad playback state matching to the periods in the content timeline.
+    ImmutableMap<Object, AdPlaybackState> adPlaybackStates =
         FakeTimeline.createMultiPeriodAdTimeline(
-            "windowId",
-            /* numberOfPlayedAds= */ 0,
-            /* isAdPeriodFlags...= */ false,
-            true,
-            true,
-            false);
+                "windowId",
+                /* numberOfPlayedAds= */ 0,
+                /* isAdPeriodFlags...= */ false,
+                true,
+                true,
+                false)
+            .getAdPlaybackStates(/* windowIndex= */ 0);
     Listener listener = mock(Listener.class);
     ExoPlayer player = new TestExoPlayerBuilder(context).build();
     player.addListener(listener);
     AtomicReference<ServerSideAdInsertionMediaSource> sourceReference = new AtomicReference<>();
     sourceReference.set(
         new ServerSideAdInsertionMediaSource(
-            new FakeMediaSource(adTimeline),
+            new FakeMediaSource(fakeContentTimeline),
             contentTimeline -> {
-              sourceReference
-                  .get()
-                  .setAdPlaybackStates(adTimeline.getAdPlaybackStates(/* windowIndex= */ 0));
+              sourceReference.get().setAdPlaybackStates(adPlaybackStates, contentTimeline);
               return true;
             }));
     player.setMediaSource(sourceReference.get());
@@ -5270,25 +5316,40 @@ public final class ExoPlayerTest {
     ArgumentCaptor<PositionInfo> newPositionArgumentCaptor =
         ArgumentCaptor.forClass(PositionInfo.class);
     ArgumentCaptor<Integer> reasonArgumentCaptor = ArgumentCaptor.forClass(Integer.class);
-    FakeTimeline adTimeline =
+    // Create a multi-period timeline without ads.
+    FakeTimeline fakeContentTimeline =
+        new FakeTimeline(
+            new FakeTimeline.TimelineWindowDefinition(
+                /* periodCount= */ 4,
+                "windowId",
+                /* isSeekable= */ true,
+                /* isDynamic= */ false,
+                /* isLive= */ false,
+                /* isPlaceholder= */ false,
+                /* durationUs= */ DEFAULT_WINDOW_DURATION_US,
+                /* defaultPositionUs= */ 0,
+                /* windowOffsetInFirstPeriodUs= */ DEFAULT_WINDOW_OFFSET_IN_FIRST_PERIOD_US,
+                /* adPlaybackStates= */ ImmutableList.of(AdPlaybackState.NONE),
+                MediaItem.EMPTY));
+    // Create the ad playback state matching to the periods in the content timeline.
+    ImmutableMap<Object, AdPlaybackState> adPlaybackStates =
         FakeTimeline.createMultiPeriodAdTimeline(
-            "windowId",
-            /* numberOfPlayedAds= */ 0,
-            /* isAdPeriodFlags...= */ false,
-            true,
-            true,
-            false);
+                "windowId",
+                /* numberOfPlayedAds= */ 0,
+                /* isAdPeriodFlags...= */ false,
+                true,
+                true,
+                false)
+            .getAdPlaybackStates(/* windowIndex= */ 0);
     Listener listener = mock(Listener.class);
     ExoPlayer player = new TestExoPlayerBuilder(context).build();
     player.addListener(listener);
     AtomicReference<ServerSideAdInsertionMediaSource> sourceReference = new AtomicReference<>();
     sourceReference.set(
         new ServerSideAdInsertionMediaSource(
-            new FakeMediaSource(adTimeline),
+            new FakeMediaSource(fakeContentTimeline),
             contentTimeline -> {
-              sourceReference
-                  .get()
-                  .setAdPlaybackStates(adTimeline.getAdPlaybackStates(/* windowIndex= */ 0));
+              sourceReference.get().setAdPlaybackStates(adPlaybackStates, contentTimeline);
               return true;
             }));
     player.setMediaSource(sourceReference.get());
@@ -5351,25 +5412,40 @@ public final class ExoPlayerTest {
     ArgumentCaptor<PositionInfo> newPositionArgumentCaptor =
         ArgumentCaptor.forClass(PositionInfo.class);
     ArgumentCaptor<Integer> reasonArgumentCaptor = ArgumentCaptor.forClass(Integer.class);
-    FakeTimeline adTimeline =
+    // Create a multi-period timeline without ads.
+    FakeTimeline fakeContentTimeline =
+        new FakeTimeline(
+            new FakeTimeline.TimelineWindowDefinition(
+                /* periodCount= */ 4,
+                "windowId",
+                /* isSeekable= */ true,
+                /* isDynamic= */ false,
+                /* isLive= */ false,
+                /* isPlaceholder= */ false,
+                /* durationUs= */ DEFAULT_WINDOW_DURATION_US,
+                /* defaultPositionUs= */ 0,
+                /* windowOffsetInFirstPeriodUs= */ DEFAULT_WINDOW_OFFSET_IN_FIRST_PERIOD_US,
+                /* adPlaybackStates= */ ImmutableList.of(AdPlaybackState.NONE),
+                MediaItem.EMPTY));
+    // Create the ad playback state matching to the periods in the content timeline.
+    ImmutableMap<Object, AdPlaybackState> adPlaybackStates =
         FakeTimeline.createMultiPeriodAdTimeline(
-            "windowId",
-            /* numberOfPlayedAds= */ 2,
-            /* isAdPeriodFlags...= */ false,
-            true,
-            true,
-            false);
+                "windowId",
+                /* numberOfPlayedAds= */ 2,
+                /* isAdPeriodFlags...= */ false,
+                true,
+                true,
+                false)
+            .getAdPlaybackStates(/* windowIndex= */ 0);
     Listener listener = mock(Listener.class);
     ExoPlayer player = new TestExoPlayerBuilder(context).build();
     player.addListener(listener);
     AtomicReference<ServerSideAdInsertionMediaSource> sourceReference = new AtomicReference<>();
     sourceReference.set(
         new ServerSideAdInsertionMediaSource(
-            new FakeMediaSource(adTimeline),
+            new FakeMediaSource(fakeContentTimeline),
             contentTimeline -> {
-              sourceReference
-                  .get()
-                  .setAdPlaybackStates(adTimeline.getAdPlaybackStates(/* windowIndex= */ 0));
+              sourceReference.get().setAdPlaybackStates(adPlaybackStates, contentTimeline);
               return true;
             }));
     player.setMediaSource(sourceReference.get());
@@ -5406,29 +5482,44 @@ public final class ExoPlayerTest {
     ArgumentCaptor<PositionInfo> newPositionArgumentCaptor =
         ArgumentCaptor.forClass(PositionInfo.class);
     ArgumentCaptor<Integer> reasonArgumentCaptor = ArgumentCaptor.forClass(Integer.class);
-    FakeTimeline adTimeline =
+    // Create a multi-period timeline without ads.
+    FakeTimeline fakeContentTimeline =
+        new FakeTimeline(
+            new FakeTimeline.TimelineWindowDefinition(
+                /* periodCount= */ 8,
+                "windowId",
+                /* isSeekable= */ true,
+                /* isDynamic= */ false,
+                /* isLive= */ false,
+                /* isPlaceholder= */ false,
+                /* durationUs= */ DEFAULT_WINDOW_DURATION_US,
+                /* defaultPositionUs= */ 0,
+                /* windowOffsetInFirstPeriodUs= */ DEFAULT_WINDOW_OFFSET_IN_FIRST_PERIOD_US,
+                /* adPlaybackStates= */ ImmutableList.of(AdPlaybackState.NONE),
+                MediaItem.EMPTY));
+    // Create the ad playback state matching to the periods in the content timeline.
+    ImmutableMap<Object, AdPlaybackState> adPlaybackStates =
         FakeTimeline.createMultiPeriodAdTimeline(
-            "windowId",
-            /* numberOfPlayedAds= */ Integer.MAX_VALUE,
-            /* isAdPeriodFlags...= */ true,
-            false,
-            true,
-            true,
-            false,
-            true,
-            true,
-            true);
+                "windowId",
+                /* numberOfPlayedAds= */ Integer.MAX_VALUE,
+                /* isAdPeriodFlags...= */ true,
+                false,
+                true,
+                true,
+                false,
+                true,
+                true,
+                true)
+            .getAdPlaybackStates(/* windowIndex= */ 0);
     Listener listener = mock(Listener.class);
     ExoPlayer player = new TestExoPlayerBuilder(context).build();
     player.addListener(listener);
     AtomicReference<ServerSideAdInsertionMediaSource> sourceReference = new AtomicReference<>();
     sourceReference.set(
         new ServerSideAdInsertionMediaSource(
-            new FakeMediaSource(adTimeline, ExoPlayerTestRunner.AUDIO_FORMAT),
+            new FakeMediaSource(fakeContentTimeline, ExoPlayerTestRunner.AUDIO_FORMAT),
             contentTimeline -> {
-              sourceReference
-                  .get()
-                  .setAdPlaybackStates(adTimeline.getAdPlaybackStates(/* windowIndex= */ 0));
+              sourceReference.get().setAdPlaybackStates(adPlaybackStates, contentTimeline);
               return true;
             }));
     player.setMediaSource(sourceReference.get());
@@ -5508,9 +5599,10 @@ public final class ExoPlayerTest {
         adPlaybackState.withPlayedAd(/* adGroupIndex= */ 2, /* adIndexInAdGroup+ */ 0);
     adPlaybackState =
         adPlaybackState.withPlayedAd(/* adGroupIndex= */ 3, /* adIndexInAdGroup+ */ 0);
-    FakeTimeline adTimeline =
+    // Create a multi-period timeline without ads.
+    FakeTimeline fakeContentTimeline =
         new FakeTimeline(
-            new TimelineWindowDefinition(
+            new FakeTimeline.TimelineWindowDefinition(
                 /* periodCount= */ 1,
                 "windowId",
                 /* isSeekable= */ true,
@@ -5520,20 +5612,19 @@ public final class ExoPlayerTest {
                 /* durationUs= */ DEFAULT_WINDOW_DURATION_US,
                 /* defaultPositionUs= */ 0,
                 /* windowOffsetInFirstPeriodUs= */ DEFAULT_WINDOW_OFFSET_IN_FIRST_PERIOD_US,
-                /* adPlaybackStates= */ ImmutableList.of(adPlaybackState),
+                /* adPlaybackStates= */ ImmutableList.of(AdPlaybackState.NONE),
                 MediaItem.EMPTY));
-
+    ImmutableMap<Object, AdPlaybackState> adPlaybackStates =
+        ImmutableMap.of(/* period.uid */ new Pair<>("windowId", 0), adPlaybackState);
     Listener listener = mock(Listener.class);
     ExoPlayer player = new TestExoPlayerBuilder(context).build();
     player.addListener(listener);
     AtomicReference<ServerSideAdInsertionMediaSource> sourceReference = new AtomicReference<>();
     sourceReference.set(
         new ServerSideAdInsertionMediaSource(
-            new FakeMediaSource(adTimeline, ExoPlayerTestRunner.AUDIO_FORMAT),
+            new FakeMediaSource(fakeContentTimeline, ExoPlayerTestRunner.AUDIO_FORMAT),
             contentTimeline -> {
-              sourceReference
-                  .get()
-                  .setAdPlaybackStates(adTimeline.getAdPlaybackStates(/* windowIndex= */ 0));
+              sourceReference.get().setAdPlaybackStates(adPlaybackStates, contentTimeline);
               return true;
             }));
     player.setMediaSource(sourceReference.get());